task 104

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,19 +9,19 @@
 int main(void)
 {
 	int i;
-       long int fibonacci[98];
+float fibonacci[98];
 
 	fibonacci[0] = 1;
 	fibonacci[1] = 2;
-	printf("%ld, %ld, ", fibonacci[0], fibonacci[1]);
+	printf("%f, %f, ", fibonacci[0], fibonacci[1]);
 
 	for (i = 2; i < 98; i++)
 	{
 		fibonacci[i] = fibonacci[i - 1] + fibonacci[i - 2];
 		if (i == 97)
-			printf("%ld\n", fibonacci[i]);
+			printf("%f\n", fibonacci[i]);
 		else
-			printf("%ld, ", fibonacci[i]);
+			printf("%f, ", fibonacci[i]);
 	}
 
 	return (0);
